+++ plugin/src/main/kotlin/com/chimerapps/storageinspector/ui/ide/view/file/FilesTree.kt
@@ -4,18 +4,27 @@ import com.chimerapps.storageinspector.api.protocol.model.file.FileInfo
 import com.chimerapps.storageinspector.api.protocol.model.file.FileServerValues
 import com.chimerapps.storageinspector.ui.util.ensureMain
 import com.chimerapps.storageinspector.ui.util.enumerate
-import com.intellij.ui.treeStructure.Tree
+import com.intellij.ide.dnd.DnDManager
+import com.intellij.ide.dnd.aware.DnDAwareTree
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.ui.TreeUIHelper
+import com.intellij.util.ui.tree.TreeUtil
 import java.awt.event.MouseAdapter
 import java.awt.event.MouseEvent
+import java.io.File
 import java.util.Enumeration
 import javax.swing.tree.DefaultTreeModel
 import javax.swing.tree.TreeNode
+import javax.swing.tree.TreeSelectionModel
 
 
 /**
  * @author Nicola Verbeeck
  */
-class FilesTree(private val onViewFileTapped: (FileInfo) -> Unit) : Tree() {
+class FilesTree(
+    private val onViewFileTapped: (FileInfo) -> Unit,
+    private val putFiles: (List<File>, String) -> Unit
+) : DnDAwareTree() {
 
     init {
         val listener = object : MouseAdapter() {
@@ -30,6 +39,10 @@ class FilesTree(private val onViewFileTapped: (FileInfo) -> Unit) : Tree() {
             }
         }
         addMouseListener(listener)
+        TreeUIHelper.getInstance().installTreeSpeedSearch(this)
+        selectionModel.selectionMode = TreeSelectionModel.SINGLE_TREE_SELECTION
+        TreeUtil.installActions(this)
+        enableDnD()
     }
 
     fun buildTree(from: FileServerValues) {
@@ -39,7 +52,7 @@ class FilesTree(private val onViewFileTapped: (FileInfo) -> Unit) : Tree() {
             var parent = root
             components.forEachIndexed { index, component ->
                 //Child
-                if (index == components.size - 1) {
+                if (index == components.size - 1 && !it.isDir) {
                     parent.childNodes += FileNode(name = component, info = it, parentNode = parent)
                 } else {
                     parent = parent.childNodes.getOrPut(component) { DirectoryNode(name = component, parentNode = parent) }
@@ -49,13 +62,30 @@ class FilesTree(private val onViewFileTapped: (FileInfo) -> Unit) : Tree() {
         root.sort()
 
         ensureMain {
+            //TODO optimize so that we update the tree instead of replacing the root node
             this.model = DefaultTreeModel(root)
         }
     }
 
+    private lateinit var myDropTarget: FilesDropTarget
+
     val selectedFile: FileInfo?
         get() = (selectionPath?.lastPathComponent as? FileNode)?.info
 
+    private fun enableDnD() {
+        if (!ApplicationManager.getApplication().isHeadlessEnvironment) {
+            myDropTarget = object : FilesDropTarget(this) {
+                override fun handleDrop(files: List<File>, into: FileSystemNode) {
+                    val target = if (into is FileNode) into.parentNode else into
+                    val targetPath = target.path
+                    putFiles(files, targetPath)
+                }
+            }
+
+            val dndManager = DnDManager.getInstance()
+            dndManager.registerTarget(myDropTarget, this)
+        }
+    }
 }
 
 private fun <E : FileSystemNode> MutableList<FileSystemNode>.getOrPut(key: String, defaultValue: () -> E): E {
@@ -70,6 +100,7 @@ private fun <E : FileSystemNode> MutableList<FileSystemNode>.getOrPut(key: Strin
 sealed interface FileSystemNode : TreeNode {
     val size: Long
     val name: String
+    val path: String
 }
 
 private class DirectoryNode(
@@ -88,6 +119,16 @@ private class DirectoryNode(
             childNodes.forEach { calculatedSize += it.size }
             calculatedSize
         }
+    override val path: String = buildString {
+        if (parentNode == null) {
+            append(name)
+        }else {
+            append(parentNode.path)
+            if (!(length == 1 && toString() == "/"))
+                append('/')
+            append(name)
+        }
+    }
 
     override fun getChildAt(childIndex: Int): TreeNode = childNodes[childIndex]
 
@@ -138,6 +179,7 @@ private class FileNode(
     val info: FileInfo,
 ) : FileSystemNode {
     override val size: Long = info.size
+    override val path: String = info.path
 
     override fun getChildAt(childIndex: Int): TreeNode {
         throw IllegalArgumentException("Children not allowed")